Fix bugs and warnings

+++ ref/sha2.c
@@ -636,13 +636,13 @@ void mgf1_256(unsigned char *out, unsigned long outlen,
 
     /* While we can fit in at least another full block of SHA256 output.. */
     for (i = 0; (i+1)*SPX_SHA256_OUTPUT_BYTES <= outlen; i++) {
-        u32_to_bytes(inbuf + inlen, i);
+        u32_to_bytes(inbuf + inlen, (uint32_t)i);
         sha256(out, inbuf, inlen + 4);
         out += SPX_SHA256_OUTPUT_BYTES;
     }
     /* Until we cannot anymore, and we fill the remainder. */
     if (outlen > i*SPX_SHA256_OUTPUT_BYTES) {
-        u32_to_bytes(inbuf + inlen, i);
+        u32_to_bytes(inbuf + inlen, (uint32_t)i);
         sha256(outbuf, inbuf, inlen + 4);
         memcpy(out, outbuf, outlen - i*SPX_SHA256_OUTPUT_BYTES);
     }
@@ -662,13 +662,13 @@ void mgf1_512(unsigned char *out, unsigned long outlen,
 
     /* While we can fit in at least another full block of SHA512 output.. */
     for (i = 0; (i+1)*SPX_SHA512_OUTPUT_BYTES <= outlen; i++) {
-        u32_to_bytes(inbuf + inlen, i);
+        u32_to_bytes(inbuf + inlen, (uint32_t)i);
         sha512(out, inbuf, inlen + 4);
         out += SPX_SHA512_OUTPUT_BYTES;
     }
     /* Until we cannot anymore, and we fill the remainder. */
     if (outlen > i*SPX_SHA512_OUTPUT_BYTES) {
-        u32_to_bytes(inbuf + inlen, i);
+        u32_to_bytes(inbuf + inlen, (uint32_t)i);
         sha512(outbuf, inbuf, inlen + 4);
         memcpy(out, outbuf, outlen - i*SPX_SHA512_OUTPUT_BYTES);
     }
